molecule_vagrant/playbooks/: Multi-instance support
Now that the vagrant module can take directly the list of instances
from molecule, update the playbooks accordingly.
This should solve a bunch of issues when the module is overwriting
the Vagrantfile when bringing up each VM. Moreover this should make
things easier to debug and more robust.

Since the provision option is not VM specific, the templates are now
expecting it in the molecule.yml driver: dict.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>